created a constants, model, and util section; added assets as example for use with constants
