+++ containers/standardization/app/constants.py
@@ -1,0 +1,24 @@
+from app.utils import read_json_from_assets
+
+
+# TODO: when we create /standardize endpoint collapse these
+sample_geocode_request_data = read_json_from_assets("sample_geocode_request_data.json")
+sample_geocode_responses = read_json_from_assets("sample_geocode_responses.json")
+sample_standardize_date_of_birth_request_data = read_json_from_assets(
+    "sample_standardize_date_of_birth_request_data.json"
+)
+sample_standardize_date_of_birth_response = read_json_from_assets(
+    "sample_standardize_date_of_birth_response.json"
+)
+sample_standardize_name_request_data = read_json_from_assets(
+    "sample_standardize_name_request_data.json"
+)
+sample_standardize_name_response = read_json_from_assets(
+    "sample_standardize_name_response.json"
+)
+sample_standardize_phone_request_data = read_json_from_assets(
+    "sample_standardize_phone_request_data.json"
+)
+sample_standardize_phone_response = read_json_from_assets(
+    "sample_standardize_phone_response.json"
+)
