+++ containers/standardization/assets/sample_standardize_date_of_birth_request_data.json
@@ -1,0 +1,14 @@
+{
+  "resourceType": "Patient",
+  "id": "c53f9ad8-34c1-ce05-c0f6-7a0ea7bd8483",
+  "name": [
+    {
+      "family": "Doe",
+      "given": [
+        "Jane"
+      ],
+      "use": "official"
+    }
+  ],
+  "birthDate": "05/30/1955"
+}
